feat(Structure): inject scripts from different game objects
Previously, many of the scripts would have an implicit relationship
with other scripts on the same game object. This meant that scripts
that relied on each other would need to be applied to the same
game object.

This change now opens up the possibility of providing the scripts
on different game objects and injecting the required script into
the using script via a public parameter.