feat(Interaction): decouple interact touch from controller object
The Interact Touch script no longer needs to be on the same script
alias game object that the touch is registered for.

Instead now it's possible to provide the script alias object as
a parameter which is then used to listen for the collision events
that have been set up on that controller.

This is achieved by the Interact Touch script no longer being
responsible for the collision checks as this is now handled by the
Controller Tracker script which emits events that the Interact
Touch script subscribes to.

This means that the actual controllers handle the raw collision
events but the Interact Touch script logic is separated from it
and can be switched at runtime.

This makes it possible to have functionality swap between the left
and right controller at runtime, example:

 * The left controller on touch will highlight the left controller
 * The right controller on touch will vibrate the right controller

This is achieved by registering a script to listen to the events
from the Interact Touch script linked to the left controller and
setting up a highlight on touch event. And a script listening to
the events on the Interact Touch script linked to the right
controller and setting up the vibrate on touch event.

Then when the script alias object parameter on each of the
Interact Touch scripts is switched, the action scripts will still
be registered to the same Interact Touch script, but the Interact
Touch script will be listening to the opposite controller.